Рефакторинг пакетов backend dao и model объеденены в domain. Реализован login по паролю.
