+++ app/domain/login.go
@@ -1,11 +1,48 @@
-package dao
+package domain
 
 import (
+	"encoding/json"
 	"github.com/google/uuid"
-	"github.com/vskurikhin/otus-highload-architect-2021-03-VSkurikhin/app/model"
+	"github.com/savsgio/go-logger/v2"
 )
 
-func (l *login) Create(login *model.Login) error {
+type Login struct {
+	id       uuid.UUID
+	Username string
+	password string
+}
+
+func (l *Login) Id() uuid.UUID {
+	return l.id
+}
+
+func (l *Login) SetId(id uuid.UUID) {
+	l.id = id
+}
+
+func (l *Login) Password() string {
+	return l.password
+}
+
+func (l *Login) SetPassword(password string) {
+	l.password = password
+}
+
+func (l *Login) String() string {
+	return string(l.Marshal())
+}
+
+func (l *Login) Marshal() []byte {
+
+	login, err := json.Marshal(*l)
+	if err != nil {
+		logger.Error(err)
+		return nil
+	}
+	return login
+}
+
+func (l *login) Create(login *Login) error {
 	// Подготовить оператор для вставки данных
 	stmtIns, err := l.db.Prepare("INSERT INTO login (id, username, password) VALUES (?, ?, ?)") // ? = заполнитель
 	if err != nil {
@@ -21,7 +58,7 @@ func (l *login) Create(login *model.Login) error {
 	return nil
 }
 
-func (l *login) Read(username string) (*model.Login, error) {
+func (l *login) Read(username string) (*Login, error) {
 
 	stmtOut, err := l.db.Prepare(`
  		SELECT id, username, password FROM login WHERE username = ?
@@ -31,16 +68,12 @@ func (l *login) Read(username string) (*model.Login, error) {
 	}
 	defer func() { _ = stmtOut.Close() }() // Закрывается оператор, когда выйдете из функции
 
-	var id uuid.UUID
-	var login model.Login
-	var password string
+	var login Login
 	err = stmtOut.QueryRow(username).
-		Scan(&id, &login.Username, &password)
+		Scan(&login.id, &login.Username, &login.password)
 	if err != nil {
 		return nil, err
 	}
-	login.SetId(id)
-	login.SetPassword(password)
 
 	return &login, nil
 }
