+++ app/domain/signin.go
@@ -1,4 +1,4 @@
-package model
+package domain
 
 import "fmt"
