+++ app/security/auth_check_token.go
@@ -1,11 +1,11 @@
-package jwt
+package security
 
 import (
 	"errors"
 	sa "github.com/savsgio/atreugo/v11"
 	"github.com/valyala/fasthttp"
 	"github.com/vskurikhin/otus-highload-architect-2021-03-VSkurikhin/app/config"
-	"github.com/vskurikhin/otus-highload-architect-2021-03-VSkurikhin/app/message"
+	"github.com/vskurikhin/otus-highload-architect-2021-03-VSkurikhin/app/domain"
 	"log"
 	"regexp"
 )
@@ -39,7 +39,7 @@ func (j *JWT) AuthCheckToken(ctx *sa.RequestCtx) error {
 	jwtCookie := ctx.Request.Header.Cookie(config.ACCESS_TOKEN_COOKIE)
 
 	if len(jwtCookie) == 0 {
-		return ctx.ErrorResponse(errors.New(message.LoginRequired.String()), fasthttp.StatusForbidden)
+		return ctx.ErrorResponse(errors.New(domain.LoginRequired.String()), fasthttp.StatusForbidden)
 	}
 
 	token, err := j.ValidateToken(string(jwtCookie))
@@ -48,7 +48,7 @@ func (j *JWT) AuthCheckToken(ctx *sa.RequestCtx) error {
 	}
 
 	if !token.Valid {
-		return ctx.ErrorResponse(errors.New(message.YourSessionIsExpired.String()), fasthttp.StatusForbidden)
+		return ctx.ErrorResponse(errors.New(domain.YourSessionIsExpired.String()), fasthttp.StatusForbidden)
 	}
 
 	return ctx.Next()
