+++ app/server/handlers/api_user.go
@@ -7,8 +7,7 @@ import (
 	sa "github.com/savsgio/atreugo/v11"
 	"github.com/savsgio/go-logger/v2"
 	"github.com/valyala/fasthttp"
-	"github.com/vskurikhin/otus-highload-architect-2021-03-VSkurikhin/app/message"
-	"github.com/vskurikhin/otus-highload-architect-2021-03-VSkurikhin/app/model"
+	"github.com/vskurikhin/otus-highload-architect-2021-03-VSkurikhin/app/domain"
 	"github.com/vskurikhin/otus-highload-architect-2021-03-VSkurikhin/app/security"
 	"strconv"
 	"strings"
@@ -19,7 +18,8 @@ func (h *Handlers) List(ctx *sa.RequestCtx) error {
 	u, err := h.Server.DAO.User.ReadListAsString()
 
 	if err != nil {
-		errorCase := message.ApiMessage{
+		logger.Error(err)
+		errorCase := domain.ApiMessage{
 			Code:    fasthttp.StatusPreconditionFailed,
 			Message: err.Error(),
 		}
@@ -30,27 +30,29 @@ func (h *Handlers) List(ctx *sa.RequestCtx) error {
 
 func (h *Handlers) Signin(ctx *sa.RequestCtx) error {
 
-	var signIn model.Signin
+	var signIn domain.Signin
 	err := json.Unmarshal(ctx.PostBody(), &signIn)
 
 	if err != nil {
-		errorCase := message.ApiMessage{
+		logger.Error(err)
+		errorCase := domain.ApiMessage{
 			Code:    fasthttp.StatusPreconditionFailed,
 			Message: err.Error(),
 		}
 		return ctx.HTTPResponse(errorCase.String(), fasthttp.StatusPreconditionFailed)
 	}
 	if logger.DebugEnabled() {
-		logger.Debugf("got: %s", signIn.String())
+		logger.Debugf("got signIn: %s", signIn.String())
 	}
 	id := uuid.New()
 	password := security.HashAndSalt([]byte(signIn.Password))
-	login := model.Login{Username: signIn.Username}
+	login := domain.Login{Username: signIn.Username}
 	login.SetId(id)
 	login.SetPassword(password)
 	err = h.Server.DAO.Login.Create(&login)
 	if err != nil {
-		errorCase := message.ApiMessage{
+		logger.Error(err)
+		errorCase := domain.ApiMessage{
 			Code:    fasthttp.StatusPreconditionFailed,
 			Message: err.Error(),
 		}
@@ -59,13 +61,14 @@ func (h *Handlers) Signin(ctx *sa.RequestCtx) error {
 	age, _ := strconv.ParseInt(signIn.Age, 10, 32)
 	sex, _ := strconv.ParseInt(signIn.Age, 10, 1)
 	ins := strings.Split(signIn.Interests, "\n")
-	user := model.Create(id, signIn.Username, &signIn.Name, &signIn.Surname, int(age), int(sex), ins, &signIn.City)
+	user := domain.Create(id, signIn.Username, &signIn.Name, &signIn.Surname, int(age), int(sex), ins, &signIn.City)
 	if logger.DebugEnabled() {
-		logger.Debugf("got: %s", user.String())
+		logger.Debugf("got user: %s", user.String())
 	}
 	err = h.Server.DAO.User.Create(user)
 	if err != nil {
-		errorCase := message.ApiMessage{
+		logger.Error(err)
+		errorCase := domain.ApiMessage{
 			Code:    fasthttp.StatusPreconditionFailed,
 			Message: err.Error(),
 		}
@@ -81,30 +84,34 @@ func (h *Handlers) Signin(ctx *sa.RequestCtx) error {
 
 func (h *Handlers) Create(ctx *sa.RequestCtx) error {
 
-	var user model.User
+	var user domain.User
 	err := json.Unmarshal(ctx.PostBody(), &user)
-	user.SetId(uuid.New())
 
 	if err != nil {
-		errorCase := message.ApiMessage{
+		logger.Error(err)
+		errorCase := domain.ApiMessage{
 			Code:    fasthttp.StatusPreconditionFailed,
 			Message: err.Error(),
 		}
 		return ctx.HTTPResponse(errorCase.String(), fasthttp.StatusPreconditionFailed)
 	}
+	user.SetId(uuid.New())
+
 	if logger.DebugEnabled() {
-		logger.Debugf("got: %s", user.String())
+		logger.Debugf("got user: %s", user.String())
 	}
 	err = h.Server.DAO.User.Create(&user)
+
 	if err != nil {
-		errorCase := message.ApiMessage{
+		logger.Error(err)
+		errorCase := domain.ApiMessage{
 			Code:    fasthttp.StatusPreconditionFailed,
 			Message: err.Error(),
 		}
 		return ctx.HTTPResponse(errorCase.String(), fasthttp.StatusPreconditionFailed)
 	}
 	msg := fmt.Sprintf("created with id: %s", user.Id().String())
-	created := message.ApiMessage{
+	created := domain.ApiMessage{
 		Code:    fasthttp.StatusCreated,
 		Message: msg,
 	}
